mod by Rohit

+++ app/src/screens/admin/ShowDeliveryBoyScreen.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import Navigation from "../../components/Navigation";
 import Footer from "../../components/Footer"; 
 import ApiCustomerService from "../../services/customer/ApiCustomerService";
-
+import Swal from 'sweetalert2';
 class DeliveryBoyListScreen extends Component{
 
     constructor(props) {
@@ -32,7 +32,27 @@ class DeliveryBoyListScreen extends Component{
     }
     
     deleteDelBoy(id){
-        ApiCustomerService.deleteDelBoy(id)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              Swal.fire({
+                icon : 'success',
+                title : 'Deleted!',
+                text : 'Your file has been deleted.',
+                showConfirmButton: true,
+                confirmButtonText: 'OKAY',
+              })
+              ApiCustomerService.deleteDelBoy(id)
+            }
+          })
+        
         .then((res) => {
            window.location.reload();
         });
@@ -55,7 +75,7 @@ class DeliveryBoyListScreen extends Component{
                 </tr>
             </thead>
             <tbody>
-                <div className="container"><h5 className="nameColor1">{this.state.boys.length == 0 && this.state.message}</h5></div>
+                <div className="container"><h5 className="nameColor1">{this.state.boys.length === 0 && this.state.message}</h5></div>
                 {this.state.boys.map(
                         boy =>
                         <tr key={boy.id}>
